requirements.txt
numpy
torch
torchvision
imgaug
lmdb
six
matplotlib
tqdm
opencv_python
progress
filetype
msgpack_python
Pillow
PyYAML
tensorboard
